libs/kernel/src/domain/entity.ts
import { UniqueEntityId } from "./unique-entity-id";

const isEntity = (v: any): v is Entity<any> => {
  return v instanceof Entity;
};

export abstract class Entity<T> {
  protected readonly _id: UniqueEntityId;
  public readonly props: T;

  constructor(props: T, id?: UniqueEntityId) {
    this._id = id ?? new UniqueEntityId();
    this.props = props;
  }

  equals(object?: Entity<T>): boolean {
    if (object === null || object === undefined) {
      return false;
    }

    if (this === object) {
      return true;
    }

    /* istanbul ignore next */
    if (!isEntity(object)) {
      return false;
    }

    return this._id.equals(object._id);
  }
}
